+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 * test: immutable setter without setter on parent object error
 * test: immutable setter now works using 'wither' method
 * fix: readonly objects no longer assumed to be value objects
+* feat: `ValueObject` attribute to explicitly mark classes as value objects
 
 ## 1.8.0